back/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-datatables==0.7.2
environ==1.0
idna==3.11
mod_wsgi==5.0.2
mysqlclient==2.2.7
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.5.0
